REFACTOR: remove sequences
