+++ unet3d/predict/__init__.py
@@ -1,2 +1,1 @@
-from .predict import *
 from .volumetric import volumetric_predictions
