--- unet3d/utils/pytorch/__init__.py
@@ -1,1 +1,0 @@
-from .dataset import WholeBrainCIFTI2DenseScalarDataset
